add validation category controller

+++ src/app/controllers/CategoryController.js
@@ -1,4 +1,6 @@
 "use strict";
+const Yup = require("yup");
+
 const categorysModel = require("../models/CategoryModel");
 
 class CategoryController {
@@ -11,12 +13,28 @@ class CategoryController {
   }
 
   async show(req, res) {
-    const categoryFound = await categorysModel.findById(req.params.id);
+    const schema = Yup.string().required("Informe um id");
 
-    res.status(200).send(categoryFound);
+    if (await schema.isValid(req.params.id)) {
+      const categoryFound = await categorysModel.findById(req.params.id);
+      return res.status(200).send(categoryFound);
+    }
+
+    res.status(400).json({ error: "O id precisa ser informado" });
   }
 
   async store(req, res) {
+    const schema = Yup.object().shape({
+      title: Yup.string().required(),
+      description: Yup.string(),
+      image: Yup.string(),
+      active: Yup.boolean().required()
+    });
+
+    if (!(await schema.isValid(req.body))) {
+      return res.status(400).json({ error: "Validation fails" });
+    }
+
     const model = new categorysModel(req.body);
 
     const result = await model.save();
@@ -25,6 +43,17 @@ class CategoryController {
   }
 
   async update(req, res) {
+    const schema = Yup.object().shape({
+      title: Yup.string().required(),
+      description: Yup.string(),
+      image: Yup.string(),
+      active: Yup.boolean().required()
+    });
+
+    if (!(await schema.isValid(req.body))) {
+      return res.status(400).json({ error: "Validation fails" });
+    }
+
     await categorysModel.findByIdAndUpdate(req.params.id, { $set: req.body });
 
     let categoryFound = await categorysModel.findById(req.params.id);
@@ -33,8 +62,14 @@ class CategoryController {
   }
 
   async delete(req, res) {
-    const deleted = await categorysModel.findByIdAndRemove(req.params.id);
-    res.status(204).send(deleted);
+    const schema = Yup.string().required();
+
+    if (await schema.isValid(req.params.id)) {
+      const deleted = await categorysModel.findByIdAndRemove(req.params.id);
+      res.status(204).send(deleted);
+    }
+
+    res.status(400).json({ error: "Informe um id" });
   }
 }
